feat: Track file metadata `$atime` file access
We utilize the `Workspace`'s `file-open` event for this. We register
for the event, then update the `$atime` if we already have existing
metadata for the file.

Because we know that file access is non-mutating, we don't perform a
full reload when this happens. We manually set the `$atime` intrinsic
and then call `store` on the file object without recursive sub-storers,
as we don't need to re-parse the file contents or other fields in this
case.

This happens for any `TFile` that has a corresponding datastore entry for its
`path` and whose metadata has the `FILE_TYPE` in its `$types` array.

If the file is a Markdown or Canvas file, we trigger the persister as well.

We then trigger an `update` for the new index revision.